fix: address critical and important issues from final review
- EmbeddingBatchProcessor: re-throw retryable exceptions before wrapping catch, so Spring Retry can intercept them
- PgVectorStoreService.deleteBySource: count chunks before delete to restore HTTP 404 contract
- application.properties: remove hardcoded DB credential defaults
- VectorStoreService.search: threshold parameter changed double -> BigDecimal
- PgVectorStoreService.parseMetadata: catch JsonProcessingException instead of Exception

+++ src/main/java/com/test/rag/service/retrieval/VectorRetrievalService.java
@@ -8,6 +8,7 @@ import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.stereotype.Service;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 @Service
@@ -34,7 +35,7 @@ public class VectorRetrievalService implements RetrievalService {
         }
 
         int topK = props.getTopK();
-        double threshold = props.getMinSimilarity().doubleValue();
+        BigDecimal threshold = props.getMinSimilarity();
 
         float[] embedding = queryEmbeddingService.embed(userQuery);
         List<ScoredChunk> results = vectorStoreService.search(embedding, topK, threshold);
